[rtsan][NFC] Move assertions and diagnostics into own impl files
