+++ compiler-rt/lib/rtsan/rtsan_assertions.h
@@ -1,4 +1,4 @@
-//===--- rtsan_stack.h - Realtime Sanitizer ---------------------*- C++ -*-===//
+//===--- rtsan_assertions.h - Realtime Sanitizer ----------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,12 +6,14 @@
 //
 //===----------------------------------------------------------------------===//
 //
+// Part of the RealtimeSanitizer runtime library
+//
 //===----------------------------------------------------------------------===//
 
 #pragma once
 
-#include <sanitizer_common/sanitizer_internal_defs.h>
+#include "rtsan_context.h"
 
 namespace __rtsan {
-void PrintStackTrace(__sanitizer::uptr pc, __sanitizer::uptr bp);
+void ExpectNotRealtime(Context &context, const char *intercepted_function_name);
 } // namespace __rtsan
