+++ compiler-rt/lib/rtsan/rtsan_context.h
@@ -38,9 +38,4 @@ private:
 };
 
 Context &GetContextForThisThread();
-
-void ExpectNotRealtime(Context &context, const char *intercepted_function_name);
-void PrintDiagnostics(const char *intercepted_function_name,
-                      __sanitizer::uptr pc, __sanitizer::uptr bp);
-
 } // namespace __rtsan
